% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_fitting.R
\name{fit_alpha_model}
\alias{fit_alpha_model}
\title{Fit Alpha Using Bayesian Pareto Model with LAI and Breakpoint Corrections}
\usage{
fit_alpha_model(
  bayesian_data,
  breakpoint,
  LAI,
  prior_mean,
  prior_sd,
  stan_model_code = stan_alpha_model,
  iter = 9000,
  warmup = 6000,
  chains = 4,
  cores = 1
)
}
\arguments{
\item{bayesian_data}{A data frame of filtered tree sizes (from \code{determine_truncation_and_filter()}).}

\item{breakpoint}{The final lower log10 size threshold (from KDE peak).}

\item{LAI}{A numeric value for site-level Leaf Area Index. The function assumes your LAI value is on
a 1-10 scale and will divide by 10 to get LAI between 0 and 1 (e.g., if you have an LAI of 5 on a scale of 1-10,
enter 5, while if you have a 50 on a scale of 1-100, enter 5)
the LAI to be between 0 and 1}

\item{prior_mean}{Prior mean for the alpha parameter.}

\item{prior_sd}{Prior standard deviation for the alpha parameter.}

\item{stan_model_code}{Character string of a Stan model (default = \code{stan_alpha_model}).}

\item{iter}{Number of Stan iterations (default = 9000).}

\item{warmup}{Number of warmup iterations (default = 6000).}

\item{chains}{Number of chains (default = 4).}

\item{cores}{Number of CPU cores to use (default = 1).}

\item{bootstrap_kde_log}{A data frame of log10(size) and log10(kernel density), used for computing R2.}

\item{refresh}{Frequency of Stan progress output (default = 0).}
}
\value{
A list with:
\describe{
\item{posterior_summary}{Data frame of posterior summaries for alpha with R2.}
\item{stan_fit}{Stan fit object.}
}
}
\description{
This function fits a Pareto model to tree size data using a Bayesian approach with Stan,
accounting for reduced visibility of small trees and site-level LAI corrections. Note that running
this function will require rstan, which itself needs Rtools. Rtools is not a CRAN package and as far
as we know must be installed directly. Try this link (https://cran.r-project.org/bin/windows/Rtools/)
}
