% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbin.R
\name{logbin}
\alias{logbin}
\title{Logarithmic binning function}
\usage{
logbin(x, y = NULL, n)
}
\arguments{
\item{x}{Numeric vector of values to be binned in the x dimension (e.g., diameter).}

\item{y}{Optional numeric vector of values to be binned in the y dimension
(defaults to \code{NULL}). Typically something like biomass production.}

\item{n}{Integer, number of bins.}
}
\value{
A data frame with the following columns:
\itemize{
\item \code{bin_midpoint} Midpoint of bins on x-axis (arithmetic mean in linear space).
\item \code{bin_value} Height of bins, either counts per unit width (if \code{y=NULL})
or sum(y)/width.
\item \code{bin_count} Number of individuals in bins.
\item \code{bin_min} Lower edge of bins on x-axis.
\item \code{bin_max} Upper edge of bins on x-axis.
}
}
\description{
This function performs logarithmic binning of values along the x-axis,
optionally aggregating y-values (e.g., biomass) into the bins.
}
\details{
This implementation is adapted from the \emph{forestscaling} package
(not available on CRAN). Please cite the original package if you use this
function in research outputs.
}
\examples{
set.seed(123)
x <- runif(100, 1, 100)
y <- rpois(100, lambda = 5)
logbin(x, y, n = 5)

}
