% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_fitting.R
\name{potential_break}
\alias{potential_break}
\title{Identify Potential Breakpoint and Prepare Kernel Density Data}
\usage{
potential_break(data, n_bootstrap = 1000, bandwidth = "SJ", trim_max = 50)
}
\arguments{
\item{data}{A data frame containing individual measurements (e.g., trees).}

\item{n_bootstrap}{Number of bootstrap replicates for KDE smoothing (default = 1000).}

\item{bandwidth}{Bandwidth method passed to \code{density()} (default = "SJ").}

\item{trim_max}{Upper size limit.}
}
\value{
A list with:
\describe{
\item{potential_breakpoint}{A numeric value representing the estimated lower log10 size bound breakpoint.}
\item{bootstrap_kde_log}{A data frame with log-scaled KDE values (\code{log_x}, \code{mean_log_density}).}
\item{original_data_trimmed}{The original size vector \emph{after} trimming by \code{trim_max}. This is crucial for quantile calculation in the first function and for the final filtering in the second.}
\item{original_raw_data_df}{The original input data frame, untouched, so the second function can filter it.}
\item{trim_max_value}{The \code{trim_max} value used in this function, passed along for consistency.}
}
}
\description{
This function performs Kernel Density Estimation (KDE) and bootstrapping
to identify a potential visibility breakpoint in size-abundance data.
It returns the estimated breakpoint and the processed KDE data. The name of the
column in the data frame that holds the size information must be called "dbh".
}
