% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_fitting.R
\docType{data}
\name{stan_alpha_model}
\alias{stan_alpha_model}
\title{Stan model string for estimating alpha with LAI and breakpoint corrections}
\format{
A character string containing the full Stan model code.
}
\usage{
stan_alpha_model
}
\description{
This is the default Stan model used by \code{\link[=fit_alpha_model]{fit_alpha_model()}}. You can modify it by replacing the \code{stan_model_code} argument.
}
\keyword{datasets}
