% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ntot_fitting.R
\docType{data}
\name{stan_totaltrees_model}
\alias{stan_totaltrees_model}
\title{Stan model string for estimating Ntotal with LAI and breakpoint corrections}
\format{
A character string containing the full Stan model code.
}
\usage{
stan_totaltrees_model
}
\description{
This is the default Stan model used by \code{\link[=estimate_total_trees]{estimate_total_trees()}}. You can modify it by replacing the \code{stan_model_code} argument.
}
\keyword{datasets}
