% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_fitting.R
\name{truncate_filter}
\alias{truncate_filter}
\title{Determine Truncation Points and filter Data for Power-Law Modeling}
\usage{
truncate_filter(breakpoint_kde_results, min_size = 10)
}
\arguments{
\item{breakpoint_kde_results}{A list returned by \code{potential_break}.
It must contain \code{potential_breakpoint}, \code{bootstrap_kde_log},
\code{original_raw_data_df}, and \code{trim_max_value}.}

\item{min_size}{The mimimum tree DBH that will be in the final distribution. This is the lower bound to
the Pareto distribution, not the breakpoint for obscured vs unobscured trees. As such, this is usually approx
10 cm.}
}
\value{
A list with:
\describe{
\item{final_breakpoint}{The final determined lower log10 size bound.}
\item{bayesian_data}{Subset of original data between 10^breakpoint and 10^upper_bound.}
\item{kerneldens_logtransform}{The full log-scaled KDE data from the first function.}
}
}
\description{
This function takes the potential breakpoint and KDE data from a previous step,
performs segmented regression, determines the upper truncation point,
and filters the original dataset to the identified size range.
}
